cleaning up mixin classes and precision callback
