+++ src/sparseml/transformers/finetune/callbacks.py
@@ -1,0 +1,116 @@
+# Copyright (c) 2021 - present / Neuralmagic, Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+import logging
+import math
+
+from transformers import TrainerCallback, TrainerControl, TrainingArguments
+from transformers.trainer_callback import TrainerState
+
+import sparseml.core.session as sml
+from sparseml.core.session import callbacks
+
+
+__all__ = [
+    "DisableHalfPrecisionCallback",
+    "PostOptimCallback",
+]
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class PostOptimCallback(TrainerCallback):
+    def on_step_end(
+        self,
+        args: TrainingArguments,
+        state: TrainerState,
+        control: TrainerControl,
+        **kwargs,
+    ):
+        """
+        Event called at the end of a training step. If using gradient accumulation,
+        one training step might take several inputs.
+        """
+        super().on_step_end(args, state, control, **kwargs)
+        callbacks.optim_post_step()
+        sml.callbacks.batch_end()
+
+    def on_substep_end(
+        self,
+        args: TrainingArguments,
+        state: TrainerState,
+        control: TrainerControl,
+        **kwargs,
+    ):
+        """
+        Event called at the end of an substep during gradient accumulation.
+        """
+        super().on_substep_end(args, state, control, **kwargs)
+        callbacks.optim_post_step()
+        sml.callbacks.batch_end()
+
+
+class DisableHalfPrecisionCallback(TrainerCallback):
+    """
+    TrainerCallback for disabling FP16 training before QAT training begins
+    :param sparseml_trainer: SparseML trainer that will call back into this object
+    :param args: args to be passed to base TrainerCallback
+    :param kwargs: key word arguments to be passed to base TrainerCallback
+    """
+
+    def __init__(self, trainer, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.trainer = trainer
+        self.on_begin_called = False
+        self.quant_start_epoch = math.inf
+
+    def check_disable(self, epoch: float, force: bool = False):
+        if (
+            force or hasattr(self.trainer, "scaler") and self.trainer.scaler._enabled
+        ) and self.qat_active(epoch):
+            self.disable_amp(epoch)
+
+    def qat_active(self, epoch: float) -> bool:
+        session = sml.active_session()
+        return session.state.model.qat_active()
+
+    def disable_amp(self, epoch: float):
+        if not self.on_begin_called:
+            # disable if training loops haven't started so we don't load
+            # the empty scaler state dict and instead disable it from the start
+            self.trainer.use_cuda_amp = False
+
+        if hasattr(self.trainer, "scaler"):
+            self.trainer.scaler._enabled = False
+
+        self.quant_start_epoch = epoch
+        _LOGGER.info(f"entering QAT phase at epoch {epoch}, disabling FP16 training")
+
+    def on_epoch_begin(
+        self,
+        args: TrainingArguments,
+        state: TrainerState,
+        control: TrainerControl,
+        **kwargs,
+    ):
+        """
+        Event called at the beginning of an epoch. Disables
+        """
+        super().on_epoch_begin(args, state, control, **kwargs)
+        self.on_begin_called = True
+        self.check_disable(state.epoch)
+
+        if state.epoch > self.quant_start_epoch:
+            _LOGGER.info(self.trainer.model)
